Move `membership_{begin,end}_date` to `lib.user.info`
